Core: Add lazy settransform
